change to byteIO, break test file into multiple
