+++ tests/tasks/__init__.py
@@ -4,7 +4,8 @@ from typing import Any, Dict, List
 from sqlalchemy.sql.ddl import CreateSchema
 
 from app.application import ContextEngine, db
-from app.crud import assets, datasets, versions
+from app.crud import assets, datasets, tasks, versions
+from app.models.enum.assets import AssetStatus, AssetType
 from app.models.orm.assets import Asset as AssetORM
 from app.settings.globals import READER_USERNAME
 
@@ -83,3 +84,75 @@ async def create_asset(dataset, version, asset_type, asset_uri, input_data) -> A
             dataset, version, asset_type=asset_type, asset_uri=asset_uri,
         )
     return new_asset
+
+
+def assert_fields(field_list, field_schema):
+    count = 0
+    for field in field_list:
+        for schema in field_schema:
+            if (
+                field["field_name_"] == schema["field_name"]
+                and field["field_type"] == schema["field_type"]
+            ):
+                count += 1
+        if field["field_name_"] in ["geom", "geom_wm", "gfw_geojson", "gfw_bbox"]:
+            assert not field["is_filter"]
+            assert not field["is_feature_info"]
+        else:
+            assert field["is_filter"]
+            assert field["is_feature_info"]
+    assert count == len(field_schema)
+
+
+async def check_version_status(dataset, version, log_count):
+    row = await versions.get_version(dataset, version)
+
+    assert row.status == "saved"
+
+    print(f"TABLE SOURCE VERSION LOGS: {row.change_log}")
+    assert len(row.change_log) == log_count
+    assert row.change_log[0]["message"] == "Successfully scheduled batch jobs"
+
+
+async def check_asset_status(dataset, version, nb_assets):
+    rows = await assets.get_assets(dataset, version)
+    assert len(rows) == 2
+
+    # in this test we don't set the final asset status to saved or failed
+    assert rows[0].status == "saved"
+    assert rows[0].is_default is True
+
+    # in this test we only see the logs from background task, not from batch jobs
+    print(f"TABLE SOURCE ASSET LOGS: {rows[0].change_log}")
+    assert len(rows[0].change_log) == nb_assets * 2
+
+
+async def check_task_status(asset_id, nb_jobs, last_job_name):
+    rows = await tasks.get_tasks(asset_id)
+    assert len(rows) == nb_jobs
+
+    for row in rows:
+        # in this test we don't set the final asset status to saved or failed
+        assert row.status == "pending"
+    # in this test we only see the logs from background task, not from batch jobs
+    assert rows[-1].change_log[0]["message"] == (f"Scheduled job {last_job_name}")
+
+
+async def check_dynamic_vector_tile_cache_status(dataset, version):
+    rows = await assets.get_assets(dataset, version)
+    asset_row = rows[0]
+
+    # SHP files have one additional attribute (fid)
+    if asset_row.version == "v1.1.0":
+        assert len(asset_row.fields) == 10
+    else:
+        assert len(asset_row.fields) == 9
+
+    rows = await assets.get_assets(dataset, version)
+    v = await versions.get_version(dataset, version)
+    print(v.change_log)
+
+    assert len(rows) == 2
+    assert rows[0].asset_type == AssetType.geo_database_table
+    assert rows[1].asset_type == AssetType.dynamic_vector_tile_cache
+    assert rows[1].status == AssetStatus.saved
